backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.1
pydantic-settings==2.9.1
sqlalchemy==2.0.40
aiosqlite==0.21.0
playwright==1.52.0
google-genai==1.14.0
httpx==0.28.1
markdown==3.8
